requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
Django==5.0.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
idna==3.6
Jinja2==3.1.3
langcodes==3.3.0
MarkupSafe==2.1.4
murmurhash==1.0.10
numpy==1.26.3
packaging==23.2
pillow==10.2.0
preshed==3.0.9
pydantic==2.5.3
pydantic_core==2.14.6
pytesseract==0.3.10
requests==2.31.0
setuptools==69.0.3
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
thinc==8.2.2
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
